+++ example/lib/src/home.dart
@@ -108,8 +108,7 @@ class _HomePageState extends State<HomePage> {
   }
 
   Widget _buildMenuBar(BuildContext context) {
-    final headerStyle = TextStyle(
-        fontSize: 11, color: Colors.grey.shade500, fontWeight: FontWeight.bold);
+    final headerStyle = TextStyle(fontSize: 11, color: Colors.grey.shade500, fontWeight: FontWeight.bold);
     final itemStyle = TextStyle(color: Colors.white);
     return ListView(
       children: [
@@ -179,10 +178,10 @@ class _HomePageState extends State<HomePage> {
     );
   }
 
-  void _launchUrl(String url) async {
-    final result = await canLaunch(url);
+  void _launchUrl(Uri url) async {
+    final result = await canLaunchUrl(url);
     if (result) {
-      await launch(url);
+      await launchUrl(url);
     }
   }